clean: re range code, fix typo and move all types to `lib/types.ts` of `storage`
